Refactor into frontend and backend components
