+++ examples/app.py
@@ -3,8 +3,8 @@ Basic example of authenticating users with OAuth2 using Microsoft's MSAL library
 Closely based on this Azure sample for Flask https://github.com/Azure-Samples/ms-identity-python-webapp
 """
 
-from typing import Union, Dict, Any
-from fastapi import Depends, FastAPI, Request
+from typing import Union
+from fastapi import Depends, FastAPI
 from fastapi.openapi.docs import get_redoc_html, get_swagger_ui_html
 from fastapi.openapi.utils import get_openapi
 from fastapi.responses import HTMLResponse, JSONResponse
@@ -23,16 +23,19 @@ app = FastAPI(
 
 # Add session middleware and authentication routes
 fastapimsal.init_auth(app)
-logged_in = fastapimsal.f_logged_in(validate=False)
-# Set auto_error=False to not raise a validation error and token will return None
-token_verified = fastapimsal.TokenVerifier(auto_error=True)
+
+# Get a callable which checks a user is logged in and raises an HTTPException otherwise
+# If we set auto_error=False the exception isn't raised and it returns None
+user_authenticated = fastapimsal.frontend.UserAuthenticated(auto_error=True)
+user_authenticated_no_error = fastapimsal.frontend.UserAuthenticated(auto_error=False)
 
 
 # Add home pages
 @app.get("/", include_in_schema=False)
-async def home(request: Request) -> HTMLResponse:
+async def home(
+    user: fastapimsal.UserIdentity = Depends(user_authenticated_no_error),
+) -> HTMLResponse:
 
-    user = request.session.get("user", None)
     if not user:
         return HTMLResponse('<a href="/login">login</a>')
 
@@ -41,10 +44,10 @@ async def home(request: Request) -> HTMLResponse:
     )
 
 
-# Place docs behind auth
+# Place docs behind auth which will error
 @app.get("/openapi.json", include_in_schema=False)
 async def get_open_api_endpoint(
-    _: dict = Depends(logged_in),
+    _: fastapimsal.UserIdentity = Depends(user_authenticated),
 ) -> Union[JSONResponse, HTMLResponse]:
     """
     Serves OpenAPI endpoints
@@ -55,7 +58,9 @@ async def get_open_api_endpoint(
 
 
 @app.get("/docs", include_in_schema=False)
-async def get_documentation(_: dict = Depends(logged_in)) -> HTMLResponse:
+async def get_documentation(
+    _: fastapimsal.UserIdentity = Depends(user_authenticated),
+) -> HTMLResponse:
     """
     Serves swagger API docs
     """
@@ -64,7 +69,7 @@ async def get_documentation(_: dict = Depends(logged_in)) -> HTMLResponse:
 
 @app.get("/redoc", include_in_schema=False)
 async def get_redocumentation(
-    _: dict = Depends(logged_in),
+    _: fastapimsal.UserIdentity = Depends(user_authenticated),
 ) -> HTMLResponse:
     """
     Serves redoc API docs
@@ -72,11 +77,11 @@ async def get_redocumentation(
     return get_redoc_html(openapi_url="/openapi.json", title="docs")
 
 
-@app.get("/apipath")
-def callme(user: Dict[str, Any] = Depends(token_verified)) -> str:
-    """An example API route"""
+# @app.get("/apipath")
+# def callme(user: Dict[str, Any] = Depends(token_verified)) -> str:
+#     """An example API route"""
 
-    if user:
-        return f"Welcome {user['preferred_username']}"
+#     if user:
+#         return f"Welcome {user['preferred_username']}"
 
-    return "Welcome - you aren't authorized. You only got here because auto_error was set to False"
+#     return "Welcome - you aren't authorized. You only got here because auto_error was set to False"
