+++ fastapimsal/frontend/__init__.py
@@ -1,0 +1,4 @@
+from .authentication import (
+    UserAuthenticated,
+    UserAuthenticatedToken,
+)
