+++ fastapimsal/init_auth.py
@@ -5,9 +5,13 @@ from fastapi.responses import RedirectResponse, Response
 from starlette.middleware.sessions import SessionMiddleware
 
 from .auth_routes import create_auth_router
-from .security import RequiresLoginException
 from .config import get_auth_settings
-from .types import LoadCacheCallable, SaveCacheCallable, RemoveCacheCallable
+from .types import (
+    LoadCacheCallable,
+    SaveCacheCallable,
+    RemoveCacheCallable,
+    RequiresLoginException,
+)
 
 AUTH_SETTINGS = get_auth_settings()
 
@@ -36,7 +40,7 @@ def init_auth(
     f_save_cache: SaveCacheCallable = default_save_cache,
     f_remove_cache: RemoveCacheCallable = default_remove_cache,
 ) -> None:
-    """Initialise the auth
+    """Add routes for Microsoft identity platform and session middleware
 
     Args:
         app (FastAPI): [description]
@@ -63,5 +67,5 @@ def init_auth(
     async def exception_handler(
         request: Request, _: RequiresLoginException
     ) -> Response:
-        "An exception with redirects to login"
+        "Redirect to homepage if login fails"
         return RedirectResponse(url=request.url_for(home_name))
